+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changes to Calva.
 ## [Unreleased]
 
 - [Always include the project creation commands in the REPL menu](https://github.com/BetterThanTomorrow/calva/issues/2575)
+- [Automatically reveal the configured result output definition on REPL connect](https://github.com/BetterThanTomorrow/calva/issues/2576)
 
 ## [2.0.457] - 2024-06-20